Revert to duplicate code for test coverage
